Refactor: rename TodoList to InsuranceList.
