+++ src/app/components/index.ts
@@ -1,4 +1,4 @@
 export { Header } from './Header';
-export { TodoList } from './TodoList';
 export { TodoItem } from './TodoItem';
 export { TodoTextInput } from './TodoTextInput';
+export { InsuranceList } from './InsurancesList';
